+++ decidim-proposals/app/controllers/decidim/proposals/admin/proposals_controller.rb
@@ -5,6 +5,7 @@ module Decidim
     module Admin
       # This controller allows admins to manage proposals in a participatory process.
       class ProposalsController < Admin::ApplicationController
+        include Decidim::Admin::Paginable
         include Decidim::ApplicationHelper
 
         helper Proposals::ApplicationHelper
@@ -101,7 +102,7 @@ module Decidim
         end
 
         def proposals
-          @proposals ||= query.result.page(params[:page]).per(15)
+          @proposals ||= paginate(query.result)
         end
 
         def proposal