+++ decidim-proposals/spec/system/admin_manages_proposals_spec.rb
@@ -23,4 +23,10 @@ describe "Admin manages proposals", type: :system do
   it_behaves_like "manage proposals permissions"
   it_behaves_like "merge proposals"
   it_behaves_like "split proposals"
+
+  context "when paginating" do
+    let!(:collection) { create_list(:proposal, 50, component: current_component) }
+
+    it_behaves_like "a paginated collection"
+  end
 end